chore: make tracing available
